Small fixes
I had an issue where the phone would swipe while the 3-dot menu was probably open, resulting in swiping up/showing the comments section and thus breaking the script. I have fixed this by also adding a delay in between the menu click and download click.

Also, the subprocess.run would return a bad return code for some reason, so I rewrote it to use the check_output, which works well for me.